Changed Dice API for discordId input
